requirements.txt
bottle==0.13.4
proxy_tools==0.1.0
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-UniformTypeIdentifiers==12.0
pyobjc-framework-WebKit==12.0
pywebview==6.1
typing_extensions==4.15.0
